+++ spec/spec_helper.rb
@@ -6,6 +6,8 @@ require "sidekiq/testing"
 Dir[File.join(__dir__, "support/**/*.rb")].sort.each { |f| require f }
 
 RSpec.configure do |config|
+  config.include Sidekiq::Rescue::RSpec::Matchers
+
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"